Refactor env attribute package structure
