+++ api/src/main/java/com/github/retrooper/packetevents/protocol/world/attributes/AttributeType.java
@@ -20,6 +20,7 @@ package com.github.retrooper.packetevents.protocol.world.attributes;
 
 import com.github.retrooper.packetevents.protocol.mapper.MappedEntity;
 import com.github.retrooper.packetevents.protocol.util.NbtCodec;
+import com.github.retrooper.packetevents.protocol.world.attributes.modifiers.AttributeModifier;
 import org.jspecify.annotations.NullMarked;
 
 /**
