+++ api/src/main/java/com/github/retrooper/packetevents/protocol/world/attributes/easing/EasingType.java
@@ -1,4 +1,4 @@
-package com.github.retrooper.packetevents.protocol.world.attributes;
+package com.github.retrooper.packetevents.protocol.world.attributes.easing;
 
 import com.github.retrooper.packetevents.protocol.mapper.MappedEntity;
 import com.github.retrooper.packetevents.protocol.util.NbtCodec;
