+++ api/src/main/java/com/github/retrooper/packetevents/protocol/world/attributes/modifiers/AlphaColorModifier.java
@@ -16,50 +16,51 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.github.retrooper.packetevents.protocol.world.attributes;
+package com.github.retrooper.packetevents.protocol.world.attributes.modifiers;
 
 import com.github.retrooper.packetevents.protocol.color.AlphaColor;
 import com.github.retrooper.packetevents.protocol.color.Color;
 import com.github.retrooper.packetevents.protocol.util.NbtCodec;
 import com.github.retrooper.packetevents.protocol.util.NbtCodecs;
+import com.github.retrooper.packetevents.protocol.world.attributes.EnvironmentAttribute;
 import org.jspecify.annotations.NullMarked;
 
 /**
  * @versions 1.21.11+
  */
 @NullMarked
-public interface ColorModifier<A> extends AttributeModifier<Color, A> {
+public interface AlphaColorModifier<A> extends AttributeModifier<AlphaColor, A> {
 
-    ColorModifier<AlphaColor> ALPHA_BLEND = (ArgbModifier) Color::blendWith;
-    ColorModifier<Color> ADD = (RgbModifier) Color::plus;
-    ColorModifier<Color> SUBTRACT = (RgbModifier) Color::minus;
-    ColorModifier<Color> MULTIPLY = (RgbModifier) Color::times;
-    ColorModifier<BlendToGrayArgument> BLEND_TO_GRAY = new ColorModifier<BlendToGrayArgument>() {
+    AlphaColorModifier<AlphaColor> ALPHA_BLEND = (ArgbModifier) Color::blendWith;
+    AlphaColorModifier<Color> ADD = (RgbModifier) AlphaColor::plus;
+    AlphaColorModifier<Color> SUBTRACT = (RgbModifier) AlphaColor::minus;
+    AlphaColorModifier<AlphaColor> MULTIPLY = (ArgbModifier) AlphaColor::times;
+    AlphaColorModifier<BlendToGrayArgument> BLEND_TO_GRAY = new AlphaColorModifier<BlendToGrayArgument>() {
         @Override
-        public Color apply(Color value, BlendToGrayArgument arg) {
+        public AlphaColor apply(AlphaColor value, BlendToGrayArgument arg) {
             return arg.blend(value);
         }
 
         @Override
-        public NbtCodec<BlendToGrayArgument> argumentCodec(EnvironmentAttribute<Color> attribute) {
+        public NbtCodec<BlendToGrayArgument> argumentCodec(EnvironmentAttribute<AlphaColor> attribute) {
             return BlendToGrayArgument.CODEC;
         }
     };
 
     @FunctionalInterface
-    interface ArgbModifier extends ColorModifier<AlphaColor> {
+    interface ArgbModifier extends AlphaColorModifier<AlphaColor> {
 
         @Override
-        default NbtCodec<AlphaColor> argumentCodec(EnvironmentAttribute<Color> attribute) {
+        default NbtCodec<AlphaColor> argumentCodec(EnvironmentAttribute<AlphaColor> attribute) {
             return NbtCodecs.ARGB_COLOR;
         }
     }
 
     @FunctionalInterface
-    interface RgbModifier extends ColorModifier<Color> {
+    interface RgbModifier extends AlphaColorModifier<Color> {
 
         @Override
-        default NbtCodec<Color> argumentCodec(EnvironmentAttribute<Color> attribute) {
+        default NbtCodec<Color> argumentCodec(EnvironmentAttribute<AlphaColor> attribute) {
             return NbtCodecs.RGB_COLOR;
         }
     }
