+++ api/src/main/java/com/github/retrooper/packetevents/protocol/world/attributes/timelines/Timeline.java
@@ -1,8 +1,9 @@
-package com.github.retrooper.packetevents.protocol.world.attributes;
+package com.github.retrooper.packetevents.protocol.world.attributes.timelines;
 
 import com.github.retrooper.packetevents.protocol.mapper.CopyableEntity;
 import com.github.retrooper.packetevents.protocol.mapper.DeepComparableEntity;
 import com.github.retrooper.packetevents.protocol.mapper.MappedEntity;
+import com.github.retrooper.packetevents.protocol.world.attributes.EnvironmentAttribute;
 import org.jspecify.annotations.NullMarked;
 
 import java.util.Map;
